testing: report to stderr instead of stdout
The testing package documentation (https://golang.org/pkg/testing/#T) says
that logs "...are accumulated during execution and dumped to standard error
when done."

Based on that we wrote some test runner automation which parses test
output watching for errors.  It does not work as expected and debugging
shows test failure reports are coming on stdout instead of stderr.
Looking through git history this output to stdout appears to have always
been the case, despite the documented behavior being stderr.

This patch makes a simple one line change to the report() function's
helper flushToParent()'s io.Writer interface to use stderr instead of
stdout.

This change though will break other users who similarly parse the output
of the testing package, but build their parsing to track stdout based
on observation that the output is actually on stdout instead of stderr.
An alternative would be to change the documentation, but the documentation
does follow the natural pattern of having a distinction between normal
outputs to stdout and error outputs to stderr.

Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>